requirements.txt
flask
flask_api
straight.plugin